+++ great_expectations/checkpoint/actions.py
@@ -7,6 +7,8 @@ The only requirement from an action is for it to have a take_action method.
 import logging
 import warnings
 
+from great_expectations.data_context.types.refs import GeCloudResourceRef
+
 try:
     import pypd
 except ImportError:
@@ -21,6 +23,7 @@ from great_expectations.checkpoint.util import (
 )
 from great_expectations.data_context.store.metric_store import MetricStore
 from great_expectations.data_context.types.resource_identifiers import (
+    GeCloudIdentifier,
     ValidationResultIdentifier,
 )
 from great_expectations.data_context.util import instantiate_class_from_config
@@ -45,6 +48,8 @@ class ValidationAction:
         validation_result_suite,
         validation_result_suite_identifier: ValidationResultIdentifier,
         data_asset,
+        expectation_suite_identifier=None,
+        checkpoint_identifier=None,
         **kwargs,
     ):
         """
@@ -56,14 +61,21 @@ class ValidationAction:
         :return:
         """
         return self._run(
-            validation_result_suite,
-            validation_result_suite_identifier,
-            data_asset,
+            validation_result_suite=validation_result_suite,
+            validation_result_suite_identifier=validation_result_suite_identifier,
+            data_asset=data_asset,
+            expectation_suite_identifier=expectation_suite_identifier,
+            checkpoint_identifier=checkpoint_identifier,
             **kwargs,
         )
 
     def _run(
-        self, validation_result_suite, validation_result_suite_identifier, data_asset
+        self,
+        validation_result_suite,
+        validation_result_suite_identifier,
+        data_asset,
+        expectation_suite_identifier=None,
+        checkpoint_identifier=None,
     ):
         return NotImplementedError
 
@@ -76,7 +88,12 @@ class NoOpAction(ValidationAction):
         super().__init__(data_context)
 
     def _run(
-        self, validation_result_suite, validation_result_suite_identifier, data_asset
+        self,
+        validation_result_suite,
+        validation_result_suite_identifier,
+        data_asset,
+        expectation_suite_identifier=None,
+        checkpoint_identifier=None,
     ):
         print("Happily doing nothing")
 
@@ -151,6 +168,8 @@ class SlackNotificationAction(ValidationAction):
         validation_result_suite_identifier,
         data_asset=None,
         payload=None,
+        expectation_suite_identifier=None,
+        checkpoint_identifier=None,
     ):
         logger.debug("SlackNotificationAction.run")
 
@@ -243,6 +262,8 @@ class PagerdutyAlertAction(ValidationAction):
         validation_result_suite_identifier,
         data_asset=None,
         payload=None,
+        expectation_suite_identifier=None,
+        checkpoint_identifier=None,
     ):
         logger.debug("PagerdutyAlertAction.run")
 
@@ -357,6 +378,8 @@ class MicrosoftTeamsNotificationAction(ValidationAction):
         validation_result_suite_identifier,
         data_asset=None,
         payload=None,
+        expectation_suite_identifier=None,
+        checkpoint_identifier=None,
     ):
         logger.debug("MicrosoftTeamsNotificationAction.run")
 
@@ -465,6 +488,8 @@ class OpsgenieAlertAction(ValidationAction):
         validation_result_suite_identifier,
         data_asset=None,
         payload=None,
+        expectation_suite_identifier=None,
+        checkpoint_identifier=None,
     ):
         logger.debug("OpsgenieAlertAction.run")
 
@@ -616,6 +641,8 @@ class EmailAction(ValidationAction):
         validation_result_suite_identifier,
         data_asset=None,
         payload=None,
+        expectation_suite_identifier=None,
+        checkpoint_identifier=None,
     ):
         logger.debug("EmailAction.run")
 
@@ -709,6 +736,8 @@ class StoreValidationResultAction(ValidationAction):
         validation_result_suite_identifier,
         data_asset,
         payload=None,
+        expectation_suite_identifier=None,
+        checkpoint_identifier=None,
     ):
         logger.debug("StoreValidationResultAction.run")
 
@@ -716,17 +745,28 @@ class StoreValidationResultAction(ValidationAction):
             return
 
         if not isinstance(
-            validation_result_suite_identifier, ValidationResultIdentifier
+            validation_result_suite_identifier,
+            (ValidationResultIdentifier, GeCloudIdentifier),
         ):
             raise TypeError(
-                "validation_result_id must be of type ValidationResultIdentifier, not {}".format(
+                "validation_result_id must be of type ValidationResultIdentifier or GeCloudIdentifier, not {}".format(
                     type(validation_result_suite_identifier)
                 )
             )
 
-        self.target_store.set(
-            validation_result_suite_identifier, validation_result_suite
+        contract_ge_cloud_id = None
+        if self.data_context.ge_cloud_mode and checkpoint_identifier:
+            contract_ge_cloud_id = checkpoint_identifier.ge_cloud_id
+
+        return_val = self.target_store.set(
+            validation_result_suite_identifier,
+            validation_result_suite,
+            contract_id=contract_ge_cloud_id,
         )
+        if self.data_context.ge_cloud_mode:
+            return_val: GeCloudResourceRef
+            new_ge_cloud_id = return_val.ge_cloud_id
+            validation_result_suite_identifier.ge_cloud_id = new_ge_cloud_id
 
 
 class StoreEvaluationParametersAction(ValidationAction):
@@ -771,6 +811,8 @@ class StoreEvaluationParametersAction(ValidationAction):
         validation_result_suite_identifier: ValidationResultIdentifier,
         data_asset,
         payload=None,
+        expectation_suite_identifier=None,
+        checkpoint_identifier=None,
     ):
         logger.debug("StoreEvaluationParametersAction.run")
 
@@ -778,10 +820,11 @@ class StoreEvaluationParametersAction(ValidationAction):
             return
 
         if not isinstance(
-            validation_result_suite_identifier, ValidationResultIdentifier
+            validation_result_suite_identifier,
+            (ValidationResultIdentifier, GeCloudIdentifier),
         ):
             raise TypeError(
-                "validation_result_id must be of type ValidationResultIdentifier, not {}".format(
+                "validation_result_id must be of type ValidationResultIdentifier or GeCloudIdentifier, not {}".format(
                     type(validation_result_suite_identifier)
                 )
             )
@@ -846,6 +889,8 @@ class StoreMetricsAction(ValidationAction):
         validation_result_suite_identifier,
         data_asset,
         payload=None,
+        expectation_suite_identifier=None,
+        checkpoint_identifier=None,
     ):
         logger.debug("StoreMetricsAction.run")
 
@@ -916,6 +961,8 @@ class UpdateDataDocsAction(ValidationAction):
         validation_result_suite_identifier,
         data_asset,
         payload=None,
+        expectation_suite_identifier=None,
+        checkpoint_identifier=None,
     ):
         logger.debug("UpdateDataDocsAction.run")
 
@@ -923,10 +970,11 @@ class UpdateDataDocsAction(ValidationAction):
             return
 
         if not isinstance(
-            validation_result_suite_identifier, ValidationResultIdentifier
+            validation_result_suite_identifier,
+            (ValidationResultIdentifier, GeCloudIdentifier),
         ):
             raise TypeError(
-                "validation_result_id must be of type ValidationResultIdentifier, not {}".format(
+                "validation_result_id must be of type ValidationResultIdentifier or GeCloudIdentifier, not {}".format(
                     type(validation_result_suite_identifier)
                 )
             )
@@ -937,9 +985,13 @@ class UpdateDataDocsAction(ValidationAction):
             site_names=self._site_names,
             resource_identifiers=[
                 validation_result_suite_identifier,
-                validation_result_suite_identifier.expectation_suite_identifier,
+                expectation_suite_identifier,
             ],
         )
+
+        data_docs_validation_results = {}
+        if self.data_context.ge_cloud_mode:
+            return data_docs_validation_results
 
         # get the URL for the validation result
         docs_site_urls_list = self.data_context.get_docs_sites_urls(
@@ -947,7 +999,6 @@ class UpdateDataDocsAction(ValidationAction):
             site_names=self._site_names,
         )
         # process payload
-        data_docs_validation_results = {}
         for sites in docs_site_urls_list:
             data_docs_validation_results[sites["site_name"]] = sites["site_url"]