+++ docs/changelog.md
@@ -4,6 +4,7 @@ title: Changelog
 
 ### Develop
 * [BUGFIX] Addresses issue #2993 (#3054) by using configuration over list keys when possible.
+* [DOCS] Added How to Get a Batch of Data From a Configured Datasource & fix BatchRequest limit param (#3382)
 
 ### 0.13.33
 * [FEATURE] Add optional ge_cloud_mode flag to DataContext to enable use with Great Expectations Cloud.