+++ docs_rtd/changelog.rst
@@ -7,6 +7,8 @@ Changelog
 
 develop
 -----------------
+* [FEATURE] Configurable multi-threaded checkpoint speedup (#3362)
+* [DOCS] "Deploying Great Expectations in a hosted environment without file system or CLI" (#3361)
 
 0.13.33
 -----------------